requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
httpx>=0.28.0,<1.0
beautifulsoup4>=4.12.0
pydantic-settings>=2.0.0
pandas>=2.0.0
numpy>=2.0.0
lxml>=5.0.0
yfinance>=0.2.36
defusedxml>=0.7.1,<1.0
PyJWT>=2.8.0,<3.0
tenacity>=8.2.0,<10.0
sqlalchemy[asyncio]>=2.0.0,<3.0
asyncpg>=0.30.0,<1.0
apscheduler>=3.10.0,<4.0
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0

# Optional ML dependencies (for KR-FinBERT sentiment)
torch>=2.0.0
transformers>=4.35.0
